add decision integration test
